ai-recommendation-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
kafka-python==2.0.2
joblib==1.3.2 